data class JSON to .kt (updated & still not completed)
