+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/auth/login/PostLoginRequest.kt
@@ -3,11 +3,9 @@ package com.binar.secondhand.kel2.data.api.model.auth.login
 
 import com.google.gson.annotations.SerializedName
 
-data class PostLoginResponse(
-    @SerializedName("access_token")
-    val accessToken: String,
+data class PostLoginRequest(
     @SerializedName("email")
     val email: String,
-    @SerializedName("name")
-    val name: String
+    @SerializedName("password")
+    val password: String
 )
