+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/auth/user/PutAuthRequest.kt
@@ -1,0 +1,23 @@
+package com.binar.secondhand.kel2.data.api.model.auth.user
+
+import com.google.gson.annotations.SerializedName
+
+data class PutAuthRequest(
+    @SerializedName("full_name")
+    val fullName: String,
+
+    @SerializedName("email")
+    val email: String,
+
+    @SerializedName("password")
+    val password: String,
+
+    @SerializedName("phone_number")
+    val phoneNumber: Int,
+
+    @SerializedName("address")
+    val address: String,
+
+    @SerializedName("image")
+    val image: String
+)
