+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/banner/get/GetBannerResponse.kt
@@ -1,0 +1,4 @@
+package com.binar.secondhand.kel2.data.api.model.seller.banner.get
+
+
+class GetBannerResponse : ArrayList<GetBannerResponseItem>()
