--- app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/banner/getBanner/GetBannerResponse.kt
@@ -1,4 +1,0 @@
-package com.binar.secondhand.kel2.data.api.model.seller.banner.getBanner
-
-
-class GetBannerResponse : ArrayList<GetBannerResponseItem>()
