+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/category/get/GetCategoryResponse.kt
@@ -1,0 +1,4 @@
+package com.binar.secondhand.kel2.data.api.model.seller.category.get
+
+
+class GetCategoryResponse : ArrayList<GetCategoryResponseItem>()
