--- app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/category/getCategory/GetCategoryResponse.kt
@@ -1,4 +1,0 @@
-package com.binar.secondhand.kel2.data.api.model.seller.category.getCategory
-
-
-class GetCategoryResponse : ArrayList<GetCategoryResponseItem>()
