+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/category/post/PostCategoryRequest.kt
@@ -1,0 +1,8 @@
+package com.binar.secondhand.kel2.data.api.model.seller.category.post
+
+import com.google.gson.annotations.SerializedName
+
+data class PostCategoryRequest(
+    @SerializedName("name")
+    val name: String,
+)
