+++ app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/product/get/GetProductResponse.kt
@@ -1,0 +1,4 @@
+package com.binar.secondhand.kel2.data.api.model.seller.product.get
+
+
+class GetProductResponse : ArrayList<GetProductResponseItem>()
