--- app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/product/getProduct/Category.kt
@@ -1,11 +1,0 @@
-package com.binar.secondhand.kel2.data.api.model.seller.product.getProduct
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Category(
-    @SerializedName("id")
-    val id: Int,
-    @SerializedName("name")
-    val name: String
-)
