--- app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/product/getProduct/GetProductResponse.kt
@@ -1,4 +1,0 @@
-package com.binar.secondhand.kel2.data.api.model.seller.product.getProduct
-
-
-class GetProductResponse : ArrayList<GetProductResponseItem>()
