--- app/src/main/java/com/binar/secondhand/kel2/data/api/model/seller/product/postProduct/PostProductResponse.kt
@@ -1,29 +1,0 @@
-package com.binar.secondhand.kel2.data.api.model.seller.product.postProduct
-
-
-import com.google.gson.annotations.SerializedName
-
-data class PostProductResponse(
-    @SerializedName("base_price")
-    val basePrice: Int,
-    @SerializedName("createdAt")
-    val createdAt: String,
-    @SerializedName("description")
-    val description: String,
-    @SerializedName("id")
-    val id: Int,
-    @SerializedName("image_name")
-    val imageName: String,
-    @SerializedName("image_url")
-    val imageUrl: String,
-    @SerializedName("location")
-    val location: String,
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("status")
-    val status: String,
-    @SerializedName("updatedAt")
-    val updatedAt: String,
-    @SerializedName("user_id")
-    val userId: Int
-)
