--- app/src/main/java/com/binar/secondhand/kel2/model/auth/register/PostRegisterResponse.kt
@@ -1,25 +1,0 @@
-package com.binar.secondhand.kel2.model.auth.register
-
-
-import com.google.gson.annotations.SerializedName
-
-data class PostRegisterResponse(
-    @SerializedName("address")
-    val address: String,
-    @SerializedName("createdAt")
-    val createdAt: String,
-    @SerializedName("email")
-    val email: String,
-    @SerializedName("full_name")
-    val fullName: String,
-    @SerializedName("id")
-    val id: Int,
-    @SerializedName("image_url")
-    val imageUrl: Any,
-    @SerializedName("password")
-    val password: String,
-    @SerializedName("phone_number")
-    val phoneNumber: Long,
-    @SerializedName("updatedAt")
-    val updatedAt: String
-)
